+++ filament/backend/CMakeLists.txt
@@ -72,11 +72,14 @@ if (NOT USE_EXTERNAL_GLES3)
             src/opengl/OpenGLPlatform.cpp
             include/private/backend/OpenGLPlatform.h
     )
+    if (EGL)
+        list(APPEND SRCS src/opengl/PlatformEGL.cpp)
+    endif()
     if (ANDROID)
         list(APPEND SRCS src/android/ExternalStreamManagerAndroid.cpp)
         list(APPEND SRCS src/android/ExternalTextureManagerAndroid.cpp)
         list(APPEND SRCS src/android/VirtualMachineEnv.cpp)
-        list(APPEND SRCS src/opengl/PlatformEGL.cpp)
+        list(APPEND SRCS src/opengl/PlatformEGLAndroid.cpp)
     elseif (IOS)
         list(APPEND SRCS src/opengl/PlatformCocoaTouchGL.mm)
         list(APPEND SRCS src/opengl/CocoaTouchExternalImage.mm)
