+++ filament/backend/src/opengl/PlatformEGL.cpp
@@ -9,7 +9,7 @@
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * WITHOUT WARANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
@@ -17,13 +17,8 @@
 #include "PlatformEGL.h"
 
 #include "OpenGLDriver.h"
-
-#include <jni.h>
-
-#include <assert.h>
-
-#include <string>
-#include <unordered_set>
+#include "OpenGLContext.h"
+#include "OpenGLDriverFactory.h"
 
 #include <EGL/egl.h>
 #include <EGL/eglext.h>
@@ -31,22 +26,11 @@
 #include <utils/compiler.h>
 #include <utils/Log.h>
 
-#include "android/ExternalTextureManagerAndroid.h"
-#include "android/ExternalStreamManagerAndroid.h"
-#include "android/VirtualMachineEnv.h"
+#include <string>
+#include <unordered_set>
 
-#include "OpenGLContext.h"
-#include "OpenGLDriverFactory.h"
+#include <assert.h>
 
-#include <android/api-level.h>
-#include <sys/system_properties.h>
-
-// We require filament to be built with a API 19 toolchain, before that, OpenGLES 3.0 didn't exist
-// Actually, OpenGL ES 3.0 was added to API 18, but API 19 is the better target and
-// the minimum for Jetpack at the time of this comment.
-#if __ANDROID_API__ < 19
-#   error "__ANDROID_API__ must be at least 19"
-#endif
 
 using namespace utils;
 
@@ -58,25 +42,17 @@ namespace glext {
 UTILS_PRIVATE PFNEGLCREATESYNCKHRPROC eglCreateSyncKHR = {};
 UTILS_PRIVATE PFNEGLDESTROYSYNCKHRPROC eglDestroySyncKHR = {};
 UTILS_PRIVATE PFNEGLCLIENTWAITSYNCKHRPROC eglClientWaitSyncKHR = {};
-UTILS_PRIVATE PFNEGLGETNATIVECLIENTBUFFERANDROIDPROC eglGetNativeClientBufferANDROID = {};
 UTILS_PRIVATE PFNEGLCREATEIMAGEKHRPROC eglCreateImageKHR = {};
 UTILS_PRIVATE PFNEGLDESTROYIMAGEKHRPROC eglDestroyImageKHR = {};
-UTILS_PRIVATE PFNEGLPRESENTATIONTIMEANDROIDPROC eglPresentationTimeANDROID = {};
-UTILS_PRIVATE PFNEGLGETCOMPOSITORTIMINGSUPPORTEDANDROIDPROC eglGetCompositorTimingSupportedANDROID = {};
-UTILS_PRIVATE PFNEGLGETCOMPOSITORTIMINGANDROIDPROC eglGetCompositorTimingANDROID = {};
-UTILS_PRIVATE PFNEGLGETNEXTFRAMEIDANDROIDPROC eglGetNextFrameIdANDROID = {};
-UTILS_PRIVATE PFNEGLGETFRAMETIMESTAMPSUPPORTEDANDROIDPROC eglGetFrameTimestampSupportedANDROID = {};
-UTILS_PRIVATE PFNEGLGETFRAMETIMESTAMPSANDROIDPROC eglGetFrameTimestampsANDROID = {};
 }
 using namespace glext;
 
-using EGLStream = Platform::Stream;
 
 // ---------------------------------------------------------------------------------------------
 // Utilities
 // ---------------------------------------------------------------------------------------------
 
-static void logEglError(const char* name) noexcept {
+void PlatformEGL::logEglError(const char* name) noexcept {
     const char* err;
     switch (eglGetError()) {
         case EGL_NOT_INITIALIZED:       err = "EGL_NOT_INITIALIZED";    break;
@@ -130,22 +106,9 @@ static unordered_string_set split(const char* spacedList) {
 
 // ---------------------------------------------------------------------------------------------
 
-PlatformEGL::PlatformEGL() noexcept
-        : mExternalStreamManager(ExternalStreamManagerAndroid::get()),
-          mExternalTextureManager(ExternalTextureManagerAndroid::get()) {
-}
+PlatformEGL::PlatformEGL() noexcept = default;
 
 Driver* PlatformEGL::createDriver(void* sharedContext) noexcept {
-    char scratch[PROP_VALUE_MAX + 1];
-    int length = __system_property_get("ro.build.version.release", scratch);
-    int androidVersion = length >= 0 ? atoi(scratch) : 1;
-    if (!androidVersion) {
-        mOSVersion = 1000; // if androidVersion is 0, it means "future"
-    } else {
-        length = __system_property_get("ro.build.version.sdk", scratch);
-        mOSVersion = length >= 0 ? atoi(scratch) : 1;
-    }
-
     mEGLDisplay = eglGetDisplay(EGL_DEFAULT_DISPLAY);
     assert(mEGLDisplay != EGL_NO_DISPLAY);
 
@@ -166,26 +129,6 @@ Driver* PlatformEGL::createDriver(void* sharedContext) noexcept {
 
     eglCreateImageKHR = (PFNEGLCREATEIMAGEKHRPROC) eglGetProcAddress("eglCreateImageKHR");
     eglDestroyImageKHR = (PFNEGLDESTROYIMAGEKHRPROC) eglGetProcAddress("eglDestroyImageKHR");
-
-    eglGetNativeClientBufferANDROID = (PFNEGLGETNATIVECLIENTBUFFERANDROIDPROC) eglGetProcAddress("eglGetNativeClientBufferANDROID");
-
-    if (extensions.has("EGL_ANDROID_presentation_time")) {
-        eglPresentationTimeANDROID = (PFNEGLPRESENTATIONTIMEANDROIDPROC)eglGetProcAddress(
-                "eglPresentationTimeANDROID");
-    }
-
-    if (extensions.has("EGL_ANDROID_get_frame_timestamps")) {
-        eglGetCompositorTimingSupportedANDROID = (PFNEGLGETCOMPOSITORTIMINGSUPPORTEDANDROIDPROC)eglGetProcAddress(
-                "eglGetCompositorTimingSupportedANDROID");
-        eglGetCompositorTimingANDROID = (PFNEGLGETCOMPOSITORTIMINGANDROIDPROC)eglGetProcAddress(
-                "eglGetCompositorTimingANDROID");
-        eglGetNextFrameIdANDROID = (PFNEGLGETNEXTFRAMEIDANDROIDPROC)eglGetProcAddress(
-                "eglGetNextFrameIdANDROID");
-        eglGetFrameTimestampSupportedANDROID = (PFNEGLGETFRAMETIMESTAMPSUPPORTEDANDROIDPROC)eglGetProcAddress(
-                "eglGetFrameTimestampSupportedANDROID");
-        eglGetFrameTimestampsANDROID = (PFNEGLGETFRAMETIMESTAMPSANDROIDPROC)eglGetProcAddress(
-                "eglGetFrameTimestampsANDROID");
-    }
 
     EGLint configsCount;
     EGLint configAttribs[] = {
@@ -356,11 +299,6 @@ Platform::SwapChain* PlatformEGL::createSwapChain(
         logEglError("eglSurfaceAttrib(..., EGL_SWAP_BEHAVIOR, EGL_BUFFER_DESTROYED)");
         // this is not fatal
     }
-    // activate this for recording frame timestamps
-    //if (!eglSurfaceAttrib(mEGLDisplay, sur, EGL_TIMESTAMPS_ANDROID, EGL_TRUE)) {
-    //    logEglError("eglSurfaceAttrib(..., EGL_TIMESTAMPS_ANDROID, EGL_TRUE)");
-    //    // this is not fatal
-    //}
     return (SwapChain*)sur;
 }
 
@@ -398,17 +336,6 @@ void PlatformEGL::makeCurrent(Platform::SwapChain* drawSwapChain,
     EGLSurface readSur = (EGLSurface) readSwapChain;
     if (drawSur != EGL_NO_SURFACE || readSur != EGL_NO_SURFACE) {
         makeCurrent(drawSur, readSur);
-    }
-}
-
-void PlatformEGL::setPresentationTime(int64_t presentationTimeInNanosecond) noexcept {
-    if (mCurrentDrawSurface != EGL_NO_SURFACE) {
-        if (eglPresentationTimeANDROID) {
-            eglPresentationTimeANDROID(
-                    mEGLDisplay,
-                    mCurrentDrawSurface,
-                    presentationTimeInNanosecond);
-        }
     }
 }
 
@@ -453,82 +380,6 @@ backend::FenceStatus PlatformEGL::waitFence(
     return FenceStatus::ERROR;
 }
 
-Platform::Stream* PlatformEGL::createStream(void* nativeStream) noexcept {
-    return mExternalStreamManager.acquire(static_cast<jobject>(nativeStream));
-}
-
-void PlatformEGL::destroyStream(Platform::Stream* stream) noexcept {
-    mExternalStreamManager.release(stream);
-}
-
-void PlatformEGL::attach(Stream* stream, intptr_t tname) noexcept {
-    mExternalStreamManager.attach(stream, tname);
-}
-
-void PlatformEGL::detach(Stream* stream) noexcept {
-    mExternalStreamManager.detach(stream);
-}
-
-void PlatformEGL::updateTexImage(Stream* stream, int64_t* timestamp) noexcept {
-    mExternalStreamManager.updateTexImage(stream, timestamp);
-}
-
-Platform::ExternalTexture* PlatformEGL::createExternalTextureStorage() noexcept {
-    return mExternalTextureManager.create();
-}
-
-void PlatformEGL::reallocateExternalStorage(
-        Platform::ExternalTexture* externalTexture,
-        uint32_t w, uint32_t h, backend::TextureFormat format) noexcept {
-    if (externalTexture) {
-        if ((EGLImageKHR)externalTexture->image != EGL_NO_IMAGE_KHR) {
-            eglDestroyImageKHR(mEGLDisplay, (EGLImageKHR)externalTexture->image);
-            externalTexture->image = (uintptr_t)EGL_NO_IMAGE_KHR;
-        }
-
-        mExternalTextureManager.reallocate(externalTexture, w, h, format,
-                AHARDWAREBUFFER_USAGE_GPU_COLOR_OUTPUT | AHARDWAREBUFFER_USAGE_GPU_SAMPLED_IMAGE);
-
-        auto ets = (ExternalTextureManagerAndroid::ExternalTexture*)externalTexture;
-        EGLClientBuffer clientBuffer;
-        if (ets->hardwareBuffer) {
-            clientBuffer = eglGetNativeClientBufferANDROID(ets->hardwareBuffer);
-            if (UTILS_UNLIKELY(!clientBuffer)) {
-                logEglError("eglGetNativeClientBufferANDROID");
-                return;
-            }
-        } else if (ets->clientBuffer) {
-            clientBuffer = (EGLClientBuffer)ets->clientBuffer;
-        } else {
-            // woops, reallocate failed.
-            return;
-        }
-
-        const EGLint attr[] = { EGL_NONE };
-        EGLImageKHR image = eglCreateImageKHR(mEGLDisplay,
-                EGL_NO_CONTEXT, EGL_NATIVE_BUFFER_ANDROID, clientBuffer, attr);
-        if (UTILS_UNLIKELY(!image)) {
-            logEglError("eglCreateImageKHR");
-        }
-        ets->image = (uintptr_t)image;
-    }
-}
-
-void PlatformEGL::destroyExternalTextureStorage(
-        Platform::ExternalTexture* externalTexture) noexcept {
-    if (externalTexture) {
-        mExternalTextureManager.destroy(externalTexture);
-        if ((EGLImageKHR)externalTexture->image != EGL_NO_IMAGE_KHR) {
-            eglDestroyImageKHR(mEGLDisplay, (EGLImageKHR)externalTexture->image);
-            externalTexture->image = (uintptr_t)EGL_NO_IMAGE_KHR;
-        }
-    }
-}
-
-int PlatformEGL::getOSVersion() const noexcept {
-    return mOSVersion;
-}
-
 void PlatformEGL::createExternalImageTexture(void* texture) noexcept {
     auto* t = (OpenGLDriver::GLTexture*) texture;
     glGenTextures(1, &t->gl.id);
@@ -543,49 +394,6 @@ void PlatformEGL::destroyExternalImage(void* texture) noexcept {
     glDeleteTextures(1, &t->gl.id);
 }
 
-backend::AcquiredImage PlatformEGL::transformAcquiredImage(backend::AcquiredImage source) noexcept {
-    void* const hwbuffer = source.image;
-    const backend::StreamCallback userCallback = source.callback;
-    void* const userData = source.userData;
-
-    // Convert the AHardwareBuffer to EGLImage.
-    EGLClientBuffer clientBuffer = eglGetNativeClientBufferANDROID((const AHardwareBuffer*) hwbuffer);
-    if (!clientBuffer) {
-        slog.e << "Unable to get EGLClientBuffer from AHardwareBuffer." << io::endl;
-        return {};
-    }
-    // Note that this cannot be used to stream protected video (for now) because we do not set EGL_PROTECTED_CONTENT_EXT.
-    EGLint attrs[] = { EGL_NONE, EGL_NONE };
-    EGLImageKHR eglImage = eglCreateImageKHR(mEGLDisplay, EGL_NO_CONTEXT, EGL_NATIVE_BUFFER_ANDROID, clientBuffer, attrs);
-    if (eglImage == EGL_NO_IMAGE_KHR) {
-        slog.e << "eglCreateImageKHR returned no image." << io::endl;
-        return {};
-    }
-
-    // Destroy the EGLImage before invoking the user's callback.
-    struct Closure {
-        void* image;
-        backend::StreamCallback callback;
-        void* userData;
-        EGLDisplay display;
-    };
-    Closure* closure = new Closure();
-    closure->callback = userCallback;
-    closure->image = hwbuffer;
-    closure->userData = userData;
-    closure->display = mEGLDisplay;
-    auto patchedCallback = [](void* image, void* userdata) {
-        Closure* closure = (Closure*) userdata;
-        if (eglDestroyImageKHR(closure->display, (EGLImageKHR) image) == EGL_FALSE) {
-            slog.e << "eglDestroyImageKHR failed." << io::endl;
-        }
-        closure->callback(closure->image, closure->userData);
-        delete closure;
-    };
-
-    return {eglImage, patchedCallback, closure};
-}
-
 void PlatformEGL::initializeGlExtensions() noexcept {
     unordered_string_set glExtensions;
     GLint n;
@@ -595,11 +403,6 @@ void PlatformEGL::initializeGlExtensions() noexcept {
         glExtensions.insert(extension);
     }
     ext.OES_EGL_image_external_essl3 = glExtensions.has("GL_OES_EGL_image_external_essl3");
-}
-
-// This must called when the library is loaded. We need this to get a reference to the global VM
-void JNI_OnLoad(JavaVM* vm, void* reserved) {
-    ::filament::VirtualMachineEnv::JNI_OnLoad(vm);
 }
 
 } // namespace filament
