+++ filament/backend/src/opengl/PlatformEGL.h
@@ -26,15 +26,9 @@
 
 #include "private/backend/OpenGLPlatform.h"
 
-struct AHardwareBuffer;
-
 namespace filament {
 
-class VirtualMachineEnv;
-class ExternalStreamManagerAndroid;
-class ExternalTextureManagerAndroid;
-
-class PlatformEGL final : public backend::OpenGLPlatform {
+class PlatformEGL : public backend::OpenGLPlatform {
 public:
 
     PlatformEGL() noexcept;
@@ -42,38 +36,40 @@ public:
     backend::Driver* createDriver(void* sharedContext) noexcept override;
     void terminate() noexcept override;
 
-    SwapChain* createSwapChain(void* nativewindow, uint64_t& flags) noexcept final;
-    SwapChain* createSwapChain(uint32_t width, uint32_t height, uint64_t& flags) noexcept final;
-    void destroySwapChain(SwapChain* swapChain) noexcept final;
-    void makeCurrent(SwapChain* drawSwapChain, SwapChain* readSwapChain) noexcept final;
-    void commit(SwapChain* swapChain) noexcept final;
+    SwapChain* createSwapChain(void* nativewindow, uint64_t& flags) noexcept override;
+    SwapChain* createSwapChain(uint32_t width, uint32_t height, uint64_t& flags) noexcept override;
+    void destroySwapChain(SwapChain* swapChain) noexcept override;
+    void makeCurrent(SwapChain* drawSwapChain, SwapChain* readSwapChain) noexcept override;
+    void commit(SwapChain* swapChain) noexcept override;
 
-    void setPresentationTime(int64_t presentationTimeInNanosecond) noexcept final;
+    bool canCreateFence() noexcept override { return true; }
+    Fence* createFence() noexcept override;
+    void destroyFence(Fence* fence) noexcept override;
+    backend::FenceStatus waitFence(Fence* fence, uint64_t timeout) noexcept override;
 
-    bool canCreateFence() noexcept final { return true; }
-    Fence* createFence() noexcept final;
-    void destroyFence(Fence* fence) noexcept final;
-    backend::FenceStatus waitFence(Fence* fence, uint64_t timeout) noexcept final;
+    void createExternalImageTexture(void* texture) noexcept override;
+    void destroyExternalImage(void* texture) noexcept override;
 
-    Stream* createStream(void* nativeStream) noexcept final;
-    void destroyStream(Stream* stream) noexcept final;
-    void attach(Stream* stream, intptr_t tname) noexcept final;
-    void detach(Stream* stream) noexcept final;
-    void updateTexImage(Stream* stream, int64_t* timestamp) noexcept final;
+    /* default no-op implementations... */
 
-    ExternalTexture* createExternalTextureStorage() noexcept final;
+    int getOSVersion() const noexcept override { return 0; }
+
+    void setPresentationTime(int64_t presentationTimeInNanosecond) noexcept override {}
+
+    Stream* createStream(void* nativeStream) noexcept override { return nullptr; }
+    void destroyStream(Stream* stream) noexcept override {}
+    void attach(Stream* stream, intptr_t tname) noexcept override {}
+    void detach(Stream* stream) noexcept override {}
+    void updateTexImage(Stream* stream, int64_t* timestamp) noexcept override {}
+
+    ExternalTexture* createExternalTextureStorage() noexcept override { return nullptr; }
     void reallocateExternalStorage(ExternalTexture* ets,
-            uint32_t w, uint32_t h, backend::TextureFormat format) noexcept final;
-    void destroyExternalTextureStorage(ExternalTexture* ets) noexcept final;
+            uint32_t w, uint32_t h, backend::TextureFormat format) noexcept override {}
+    void destroyExternalTextureStorage(ExternalTexture* ets) noexcept override {}
 
-    void createExternalImageTexture(void* texture) noexcept final;
-    void destroyExternalImage(void* texture) noexcept final;
+protected:
+    static void logEglError(const char* name) noexcept;
 
-    backend::AcquiredImage transformAcquiredImage(backend::AcquiredImage source) noexcept final;
-
-    int getOSVersion() const noexcept final;
-
-private:
     EGLBoolean makeCurrent(EGLSurface drawSurface, EGLSurface readSurface) noexcept;
     void initializeGlExtensions() noexcept;
 
@@ -82,17 +78,13 @@ private:
     EGLSurface mCurrentDrawSurface = EGL_NO_SURFACE;
     EGLSurface mCurrentReadSurface = EGL_NO_SURFACE;
     EGLSurface mEGLDummySurface = EGL_NO_SURFACE;
-    EGLConfig mEGLConfig;
-    EGLConfig mEGLTransparentConfig;
-    int mOSVersion;
+    EGLConfig mEGLConfig = EGL_NO_CONFIG_KHR;
+    EGLConfig mEGLTransparentConfig = EGL_NO_CONFIG_KHR;
 
     // supported extensions detected at runtime
     struct {
         bool OES_EGL_image_external_essl3 = false;
     } ext;
-
-    ExternalStreamManagerAndroid& mExternalStreamManager;
-    ExternalTextureManagerAndroid& mExternalTextureManager;
 };
 
 } // namespace filament
